+++ src/app/restaurant-reservation/page.tsx
@@ -9,23 +9,18 @@ export default function Reservation() {
   const [date, setDate] = useState<Date>(new Date())
   const a = useSearchParams()
   //const id = a.get("id")
-  const id = 1 // 변경 필수
+  const id = "1" // 변경 필수
 
   const onDateChange = (e : any)=>{
     setDate(e)
   }
 
-  const onReservation = async()=>{
+  const onReservation = async(id : string, date : Date)=>{
     const formmatDate = date.toISOString().slice(0,10)
     await fetch(`http://localhost:3001/restaurant/reservation/${id}`,{ 
       method : 'POST',
       headers : {"Content-Type" : "application/json"},
-      body : JSON.stringify({
-        date : formmatDate,
-        userId : 3, // 하드코딩 변경해야함
-        userName : "권기현", // 하드코딩 변경해야함
-        restaurantId : id
-      })
+      body : JSON.stringify({date : formmatDate})
     }).then(()=>{
       alert(`${formmatDate}에 예약성공했습니다`)
     }).catch(e=>{
@@ -39,7 +34,7 @@ export default function Reservation() {
  <div>
   <Calander onChange={onDateChange} value={date}/>
  </div>
- <button className="border border-black" onClick={onReservation}>예약하기</button>
+ <button className="border border-black" onClick={()=>onReservation(id, date)}>예약하기</button>
 </div>
   );
 }
